requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0

# Deep Learning
tensorflow>=2.10.0
keras>=2.10.0
protobuf==6.31.1

# Real-time Streaming
kafka-python>=2.0.2

# Database
psycopg2-binary>=2.9.0

# Backend API
flask>=2.3.0
flask-cors>=4.0.0
flask-restx>=0.5.0
flask-sqlalchemy>=3.0.0

# Database ORM
sqlalchemy>=2.0.0
alembic>=1.12.0

# Data Visualization
matplotlib>=3.5.0
seaborn>=0.12.0
plotly>=5.13.0

# Utilities
python-dotenv>=1.0.0
python-dateutil>=2.8.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Jupyter for notebooks
jupyter>=1.0.0
ipython>=8.0.0

# Additional utilities
joblib>=1.0.0
scipy>=1.7.0
requests>=2.28.0
